Refactored map
