+++ src/views/HelpAndKnowledge/SolutionsHelp/_components/SolutionHelpCard/index.scss
@@ -1,0 +1,11 @@
+.solution-card {
+    .usa-card__container {   
+        border-radius: .25rem;
+    }
+}
+
+.shadow {
+    .usa-card__container {   
+        box-shadow: 0px .5rem 1rem rgba(0, 0, 0, 0.1)
+    }
+}
